🔥 Folder Structure For Components is done
Works properly for simple projects
